pyproject.toml
[project]
name = "docassistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexander Kleptsov",email = "angst1437@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "docx2pdf (>=0.1.8,<0.2.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "docling (>=2.28.0,<3.0.0)",
    "spacy (>=3.8.4,<4.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "aiogram (>=3.19.0,<4.0.0)"
]

[tool.poetry]
readme = "README.md"
package-mode = false


[tool.poetry.group.dev.dependencies]
asyncio = "^3.4.3"
django = "^5.1.7"
logging = "^0.4.9.6"
ollama = "^0.4.7"
dotenv = "^0.9.9"
python-docx = "^1.1.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
